Updated openstack/openstack Project: openstack/keystone  ee2747bac2e6eeb86e6340a2bb3509bc25f115bc
Corrected punctuation on multiple exceptions

As mentioned in the bug report, keystone/exceptions.py has many
exceptions with messages that are not consistant with each other
and have various punctuational differences, such as some ending
with a period while others do not.

This change adds a '.' to the end of many exception messages and
adds small changes to other messages in order to keep all of the
exception messages consistant.

Change-Id: I21cac56ff70dbc2693c6090b887537a7c1f303e1
Closes-Bug: #1656026