Updated openstack/openstack Project: openstack/swift  aebcc5898c4b7d643b066bd6477fb7b840400e53
Fix object server tests to include content-type headers

In [1] it was necessary to override the monkey patching
of mimetools so that mimetools would apply a default
Content-Type=text/plain header when missing from object server
requests. Otherwise some tests fail due to Content-Type=None.

This patch fixes the fragile tests by adding a Content-Type header,
and thus removes the need to override the mimetool monkey patching.
This is closer to the real world in which mimetools is patched in
the object server and PUT requests should have a Content-Type.

Drive-by fix some typos and apply config assertions to all policies.

[1] Related-Change: I5b5f90bb898a335e6336f043710a05a44e3b810f

Change-Id: I4c5bca42f753be4165cb731b8e3957eb4cdd28d5