Updated openstack/openstack Project: openstack/neutron-lbaas  85f08fcfed0c7d43f38b6387b301012f7c9afdcc
Only if loadbalancer's status is normal then create pool in db

When a loadbalancer is in PENDING_UPDATE state, creating a pool with
this lb will return http 409 error with "Invalid state PENDING_UPDATE
of loadbalancer resource LB_ID". But the pool already exists in the
database although the api return this error.

So we shall create a pool in db only if the loadbalancer's state is
normal, wich is consistent with creating listners.

Change-Id: Idb35d4e6ecdc46ff3748768774737b3b01d515f6
Closes-Bug: #1650170