Updated openstack/openstack Project: openstack-dev/grenade  74ce189956730572d0ef53c1fba4352cb230a279
Fix cellsv2 cell0 database name

For some reason we were defaulting the name of the cell0 database
to nova_api_cell0 instead of nova_cell0. Devstack inherited that to
make things work, but we don't really want that. This patch makes us
use the proper name and create the cell0 mapping accordingly. As a
side effect, it also starts the process of unifying the cellsv1 and
cellsv2 paths by creating the cell0 mapping the same for both.

This is the grenade side of I4e7f6c5eaa068c98e5c4ef3feaee50d8e4f5d484.

The related Nova change which depends on this is:

I86797785f76c2c927a4db8fef72b8f8d986af6b9

We have to change the from-update/upgrade-nova script to explicitly
call map_cell0 so we can control the database connection used to
workaround the bug being fixed in that Nova change.

Related-Bug: #1656673

Change-Id: Ifdcd6a572dc0b42ba852241e22f6713cd2b084b0