Updated openstack/openstack Project: openstack/tempest  29fbb70ce04bf67e2471d4ddc873c603698e9d4c
Use Token Clients from the client factory

Token clients were still instantiated separately in clients.py.
For consistency we should use the token clients from the client
factory in service_clients.

Since token clients have a very different interface, they will
ignore most of the parameters passed to them, and they need
an auth_url to be specified, as they don't use an auth_provider
like all other clients. When token clients call __init__ on
super(), they force service, region and auth_provider to None, so
we must ensure that they are not in the kwargs passed through
to super as well.

Documenting this in the ServiceClients class docstring as well
so that consumers know how instantiate a token client from it
in case they need to.

Change-Id: I95a3fd1a8859e807e04717eae82bf01f0d9a2085