Updated openstack/openstack Project: openstack-infra/nodepool  c48dbe9a9142dca5e3ccb98d5cb27e42deaf89ca
Add dib-image-list.json webapp target

I have monitoring scripts that push the build status to me.  This
became a little harder with the split to multiple builders, as the
logs for each build are kept on the builder.

Rather than parsing tables, add a .json target to the dib-image-list
so my script can find the most recent build, see the builder and
retrieve the latest logs from there.

Future work is to do the image-list and make a nicer status page, if
this approach is OK.

Change-Id: I3410a4e4efd649732747b0502d66d212f50fa1bb