Updated openstack/openstack Project: openstack/tempest  a6d135128ee8236da3b58eb057576929008ecc6b
Use invalid id with rand string instead of special char

Many negative tests use invalid id with combination of special char
like '#$%%&^&^'.
It has the hash tag which splits the url to a url + fragment identifier part.
Some WSGI backends (such as uwsgi) might cut this part of the url,
and pass it in a different environment variable.

Change-Id: Id69baf8d364f31e6fa745929fea890652d7ee118
Closes-bug: #1545407