Updated openstack/openstack Project: openstack/tripleo-heat-templates  e0d1ab8da4950d2ae7c82992289ae60ee5fb363b
Continue checking for request url if initially null

When using get-occ-config.sh during overcloud upgrades, the script could
potentially be started before the stack has been upgraded. In that
scenario, the script will return null for the request metadata url for
the deployed-server resource since the stack has not yet been updated
and it's still using the previous Heat signaling.

This patch updates the script to just continue checking the resource
metadata if the request url is null. Eventually, once the resource has
been updated, the script will continue and properly populate the
os-collect-config configuration.

Change-Id: I9db54d8ad278715f42b768edf8f0fd21998b2098