Updated openstack/openstack Project: openstack/glance  8a25f2734079319f050e3b4e2847c268417b5d85
Adjust test suite for new psutil versions

psutil 2.x and above has a lot of API changes as described in:
https://github.com/giampaolo/psutil/blob/master/HISTORY.rst

So we should work correctly with both old and new psutil
versions by using a version check and use the correct
method/attributes. This allows to eventually unlift the version
cap that starts to hurt.

Related-Bug: 1645918
Change-Id: I6be775e83876271012f6b7a777ea2b5cc3a008f6