Updated openstack/openstack Project: openstack/neutron  9a8d7056879fa47977d3f5bb8c9a88fde6bf2606
ovsfw: Raise exception if tag cannot be found in other_config

Previously, if tag was not present in other_config obtained from ovsdb
for any reason, DEAD VLAN tag was used. This is not smart at all as it
puts all conntrack entries to one point. Also tag is mandatory and if
other_config doesn't contain it, it's a huge mistake that should never
happen.

Change-Id: I91ab75b52b70dbba4c7823550bfdfe0ab9396336
Related-bug: 1564947