Updated openstack/openstack Project: openstack/cinder  d9e48eb1a42ee13f89a7ecc9aee4d3d3709640fe
Fix response bootable parameter type in api document

This commit fixes the response bootable parameter type in the api
document. The type is string not boolean in the current code. And
the request parameter's type is boolean. I think this is an
inconsistency but we shouldn't change it to keep the backward
compatibility.

Change-Id: Ia0c70341f595580d12888e1f517e314fb98d4a6b