Updated openstack/openstack Project: openstack/keystone-specs  fd85ae9ea339c14e67b01670861d09e1e582a990
Update shadow mapping spec

As we work through the implementation for auto-provisioning, it was
noted that we should update the spec to make sure it is consistent
with the implementation. We want to clarify that roles defined in the
shadow mapping aren't created on-the-fly, like other resources are.

This is due to the fact that policy changes might be required
across the deployment in order to use the role. This could still lead
to unfriendly user experience if a user is allow to authenticate,
gets a set of projects, then can't actually do anything with them
since the policy hasn't taken the new shadow role into consideration.

This is something that we could probably change in the future if we
find a way to work around the synchronization of policy and roles.

bp shadow-mapping

Change-Id: Ifc0a7191ec6717934564851f7b907ff815da7d2d