Updated openstack/openstack Project: openstack/python-manilaclient  c91deb849839fec4f14f35e9a6271aaaaa64deec
Remove Nova Net option for share networks

Nova network was deprecated in Newton and is no longer supported for
regular deployments in Ocata [1].  In manila itself, changes [2] and
[3] remove nova network plugin support in manila services and remove
the nova-network-id field from the database model for share networks.

Add a new micro-version in which nova-network-id options are no longer
allowed when creating or modifying share networks.

[1] http://docs.openstack.org/releasenotes/nova/unreleased.html
[2] I846b760fa7c01f7f86768399a2bfad9ced7e57cd
[3] I8b9a559fbea61979f01737ed1dc272276c4f1269

Partially-implements: bp remove-nova-net-plugin

Closes-Bug: #1656756

Change-Id: Ic27b7d26a5a7d1100e57ee8a857fc45fc3b0a909