Updated openstack/openstack Project: openstack/cinder  559d63f7d445bf979cf5addb4fceb312ddb13fe4
Save model_update as admin in create_volume

admin_metadata are not saved by the object without an admin context.
Turns out Huawei driver returns admin_metadata in model_update. To make
sure admin_metadata will be saved we should save the volume to the DB
with admin context. This commit implements that.

Change-Id: Ia48498734db61474009c8854a353c67cbdfccc0e
Closes-Bug: 1656867