Updated openstack/openstack Project: openstack/ironic  0fa21ef54b73a5752523ee7b0e20ac6e7635451b
Allow to attach/detach VIF to portgroup

With this patch port groups are activated in Ironic.
When attaching a VIF to a node, it is attached to the first
free port group. If there are no free port groups, the first
available port (pxe_enabled has higher priority) is used
instead.

Related-Bug: #1618754
Related-Bug: #1582188
Co-Authored-By: Vladyslav Drok <vdrok@mirantis.com>
Change-Id: I0dca2c2d98184e370c08c3e05aa3edadead869af