Updated openstack/openstack Project: openstack/cinder  e90d1b01f63ba568013d4de4634bdbdef7bdebb4
Skip CI link in list for infra supported drivers

For lack of a better way, for drivers that have CI run by Jenkins we
denote this using the name 'Cinder_Jenkins'. CI links were added to
the generated driver list, but for these drivers this results in a
link to an invalid wiki page. For these drivers we should just skip
adding the CI info link.

Change-Id: I8a3087f8102392ca604a42fdf006e1a4e1a026d6