Updated openstack/openstack Project: openstack/ironic  10b2951c0829924d41b96d74e662b7061e5aca4e
Ensures that OneView nodes are free for use by Ironic

Following the requirement. The validation is used within
"OneViewPower.validate". As a result it enforces that a machine
is owned by OneView, thus it cannot be powered on or off by the
Ironic driver. Raising an exception when the node is owned by
Oneview.

Change-Id: Ibf500577b7a3b5976fd49fd217c1be3d3e9c0f46
Closes-Bug: 1627034
Co-Authored-By: Stenio Araujo <steniaraujo@lsd.ufcg.edu.br>