Updated openstack/openstack Project: openstack-infra/shade  1128a49e80cd38aed7f7dbaed34128279bf31c38
Add support to task manager for async tasks

In order to support replacing SwiftService, we need to be able to spin
some tasks into async threads. We'll allow for calculation of rate
limiting to match with the start of the call, but will return a Future
to the call site. This patch is followed by a patch to Nodepool to have
Nodepool's TaskManager subclass shade's TaskManager.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>

Change-Id: Ib473ceece7169d76dd37702d66b188a841ec3f59