Updated openstack/openstack Project: openstack/nova  099e0be427b533e893d6e875e9eadba602df977c
Fix the generated cell0 default database name

This makes the default database name for cell0 be based on the
main database name, not the api one. Before this, we would get
a database name of nova_api_cell0 by default, intead of nova_cell0.
This was confusing because cell0 has "main" schema inside, not "api"
schema.

Closes-Bug: #1656673
Change-Id: I86797785f76c2c927a4db8fef72b8f8d986af6b9