Updated openstack/openstack Project: openstack/oslo.middleware  472363fb2dc4997a51f75749c96eafa25d3008b5
Add new middleware to send API data to statsd

The new StatsMiddleware is a Paste filter that examines the URL path and
request method, and sends a stat count and a timer to a statsd host
whose name is based on the path/method.

If your statsd is configured to send stats to Graphite, you'll end up
with stat names of the form:
timer.<appname>.<METHOD>.<path>.<from>.<url>

Because a dot has special meaning in Graphite, dots in API versions that
appear in the path will be replaced with _, so for example v2.1 becomes v2_1,
and v1.0 becomes v1_0.

Change-Id: Ieaffeded1bf81c0782d88f49b6f5209f11744899