Updated openstack/openstack Project: openstack-infra/project-config  b23b63a70af5ba4268db5513cfb215cb9d3de9a1
Adds translation jobs to octavia

With the merger of neutron-lbaas into the octavia project, we need
to enable translation on the octavia project.
This patch updates the octavia project to have the translation jobs
and translate option.

This patch also updates the octavia project description to reflect
it's new role in OpenStack.

Change-Id: Iadcdd330e415d9ab32b7437f7b9ef887c4d5c68f