Updated openstack/openstack Project: openstack-infra/project-config  0366f25a587d45fb6a014699e5d934e0778f22a6
Add tripleo overcloud upgrade job to t-h-t repo check

This adds the new upgrade job to the check queue, only for the
tripleo-heat-templates repo initially (it's still available via
the experimental queue for other repos).

This has been proven via a number of experimental job passes, e.g
see:
  https://review.openstack.org/#/c/404831 (adds support to tripleo-ci)
  https://review.openstack.org/#/c/412475
  https://review.openstack.org/#/c/419894
  https://review.openstack.org/#/c/420027
  https://review.openstack.org/#/c/420058

I've also moved it to the experimental queue for tripleo-ci, as we
will still want to run checks on any tripleo-ci patches touching
the upgrade related code, possibly it will be needed on other repos
too in future, but it's not yet clear which ones.

Change-Id: Ib53854853700c82fae522fc52fe2e8c87a10225a