Updated openstack/openstack Project: openstack-dev/pbr  7122fd77f6df534554f6de397482e3f3db6fc415
Add Constraints support

Adding constraints support to clients is slightly more complex than
services as the clients themselves are listed in
upper-constraints.txt which leads to errors that you can't install
a specific version and a constrained version.

This change adds constraints support by also adding a helper script
to edit the constraints to remove pbr.

Change-Id: I80a10f80e2c5efafb099a6c1dcebe8f0935ace09