Updated openstack/openstack Project: openstack-infra/devstack-gate  60e957c75797bf69cf8e2c5b9beb215d85998949
functions: sort rpms list in cleanup_host

In cleanup_host, we find useful to sort the rpms by alphabetical name so
it's easier to use when people copy/paste the list with another list to
make packages diff list.

It's also easier for readability when going through the list to find a
package.

Change-Id: I9baac97f94f9dbc8d96b2d625fc1be8f4cb9c6a6