Updated openstack/openstack Project: openstack-infra/release-tools  d818b491ce8ec8adf389261cc818993cac91de96
skip reno when release-notes is only getting a list of changes

If all we want is the list of changes, such as when called from
list_unreleased_changes.sh, then do not try to scan with reno. The scan
will fail anyway, because HEAD isn't a tag and reno rejects it.

Change-Id: I681ac8612758c0cd2a3abb62b11c36191c83861c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>