Updated openstack/openstack Project: openstack-infra/project-config  7bbac129003ab34150536f90d3023bba7fe0a480
tempest-dsvm-neutron-full-ssh-master voting

Making the gate-tempest-dsvm-neutron-full-ssh-master job voting
and adding it to tempest gate. This job uses the ssh validation
in tempest, and it's currently running as experimental against
nova and neuton, in check (non-voting) for Tempest as well as
a periodic job (periodic-tempest-dsvm-neutron-full-ssh-master).

The job has been fixed at the beginning of Dec 2016 and has been
since rather stable. Making the job gating for Tempest would allow
us to gather more data about its stability.

http://status.openstack.org/openstack-health/#/g/build_name/periodic-tempest-dsvm-neutron-full-ssh-master?duration=P3M

The aim eventually is to turn on validation in
gate-tempest-dsvm-neutron-full-ubuntu-xenial.

I found one case in which the test_reboot_hard failed, as ssh did
not work after hard reboot, with no connectivity at all to the VM.
The console logs showed that the issue was related to:

modprobe: module virtio_blk not found in modules.dep

Running this job in the gate may help discovering more cases of this
and produce enough data to fix an eventual issue.

Change-Id: Ia56cca96df159e3d1605c49a2be1a3e4796a07aa