Updated openstack/openstack Project: openstack/ironic-python-agent  7680625994093f39fa9eec8aa2d8c17c35f44e81
Use sh instead of bash when installing grub

there is no guarantee that 'bash' is present in a user's image,
while 'sh' as system shell must be there.
As we do not use any bash-specific syntax (just starting commands in a
subshell) when installing the bootloader from chroot-ed user image,
let's be more supportive and use 'sh' instead of 'bash' for that.

Change-Id: I1fb82068b9c55da35166d8d2ecf9f0ba41356adb
Closes-Bug: #1657096