Updated openstack/openstack Project: openstack/ironic  f5735999c6b35538ea013aaa0ede7d678b48d6e3
Expose default interface calculation from driver_factory

The API for this feature will need to be able to calculate the default
interfaces for a given hardware interface. The code was already there to
do it, but just needed a bit of refactoring and a public method. Move
things around as we need them, and add a few tests to test the
default_interface method directly.

Change-Id: I993b76b7a2e6219cacc86558a612ecf0ba10e685
Partial-Bug: #1524745