Updated openstack/openstack Project: openstack-infra/project-config  dd8e58a9664e58c7e5f9ce582950b722a8631847
nova-dpm: enabled the job for read-the-docs (rtd)

When we created the nova-dpm project with commit e6573087, we weren't
yet quite sure if we wanted to use the publishing of the docs to
read-the-docs, so we omitted the job in the initial setup.
After some time we realised that this is the most reasonable place
for our docs and release notes. This change enables therefore this job.

Change-Id: Ieb37ec4d102e15ebb26d2d7f0ea041fdae0e3663
Closes-Bug: 1640841