Updated openstack/openstack Project: openstack-infra/tripleo-ci  9c7a2bb01a98ba3848be9e57ab8f1d105442ab01
Failure debugging improvements

A couple of small changes intended to make it easier to debug
failures in the gate.  First, run nova service-list and
neutron agent-list after failures to create the ping test stack,
not just after a failure to ping the vm.  A service outage could
also cause a stack create failure.

Second, sort the rpm -qa output for easier comparing between
different runs of a job.  Otherwise it seems to come out in a
semi-random order that doesn't diff nicely.

Change-Id: I9e3319811054668289193b24fd11ce6f0775ade9