Updated openstack/openstack Project: openstack-infra/project-config  4ecca9d47aaa8918ba1bc4d1f37917a0ea58dc03
Make tripleo ovb updates job voting

This job has been running for a while now and there have not been
any more spurious failures in it than in the other ovb jobs.  Since
it has also caught some legitimate problems in that time but was
ignored because it's non-voting, we'd like to move it to voting.

Change-Id: I3c328699394474d35ad2eb3088d4efb44829c579