Updated openstack/openstack Project: openstack/os-refresh-config  e1e0ef09166e618dd97562dd414bd04bd509062d
Adjust test suite for new psutil versions

psutil 2.x and above has a lot of API changes as described in:
https://github.com/giampaolo/psutil/blob/master/HISTORY.rst

So we should work correctly with both old and new psutil
versions by using a version check and use the correct
method/attributes. This allows to eventually unlift the version
cap that starts to hurt.

Change-Id: I6b6682e3af34f43da4e397d602e9ee4e1c83970f
Related-Bug: 1645918