Updated openstack/openstack Project: openstack/python-cinderclient  f63226ee491491caf5d8261e2959769675c00656
Metadata based snapshop filtering

The snpapshot-list API for cinder gives a list of snapshots based
on certain criteria to the user. From microversion 3.22 onwards
the snapshot-list API has been enhanced to support snapshot list
filtering based on metadata of snapshots. The metadata is stored
as key-value pair for every snapshot.
With this commit cinder will be queried based on metadata key and
value specified in the API snaphot-list. All the snapshots which
match the key, value provided by the user along with any other
filter criteria will be returned.
Added the test cases for the CLI and web requests.

DocImpact: "Filters results by a metadata key and value pair.
Default=None." on cinder snapshot-list
APIImpact

Closes-bug: #1569554

Change-Id: Idec0d0d02e7956843f202508e32c023c3cafbb0f