Updated openstack/openstack Project: openstack-infra/tripleo-ci  23d7fffa65eecaaade67c88aa235aaf58ca5eb3d
Deploy network envs appropriate for the job

The nonha jobs don't use net-iso, so there's no need to spend the
time creating a lot of networks and ports.  In addition, OVB now
has the ability to deploy a network environment that supports
basic bonding, and this change adds support for deploying that as
well.  No jobs currently use bonding, but that will be added in a
follow-up patch.

Change-Id: Ifb65d962293b8b69b2a84597c29c1ffae5d9bc2c