Updated openstack/openstack Project: openstack-infra/project-config  1b096042815f1555e9608cabe504b248779164f2
adjust the pattern for updating constraints file URLs

With more projects using tox_install.sh, fewer match the pattern
starting with "-c" that was being passed directly to pip. Remove that
part of the pattern and rely on just matching the {env:...} syntax that
tox uses to pass values to arbitrary commands.

For an example of a project where the old version of the script failed
but this change worked, see Iffd79e8d267142579f71248e8b78ecb59f07d522

Change-Id: Idc25df0100153c324674f6d60fe5cfceaea59d54
Signed-off-by: Doug Hellmann <doug@doughellmann.com>