Updated openstack/openstack Project: openstack/trove  7d219aaf59ed76db3cd10b9a15e167cab55bdce2
[gate fix] Run resize-* for MySQL only

With the advent of different clouds running tests in the gate,
resize-volume and resize-instance have been failing sporadically. With
the sheer number of jobs that trove runs, the odds of having these tests
pass has dropped to almost zero. In order to facilitate the test runs
passing, these tests have be limited to just MySQL.

They can be re-enabled once the bugs have been fixed. Note that
the issues may be in the neutron devstack setup (or something
related to xenial) and not in the trove code specifically.

Note that the way the groups are set up, instance_error
and isntance_force_delete are also relegated to MySQL only.
Since these are datastore agnostic tests, that should
not be a problem.

Change-Id: I9c6016c6fd2f37918f1ad4fee51b6f3682f1e15b
Related-Bug: 1657249