Updated openstack/openstack Project: openstack-dev/devstack  b1d3977daf6463854b44fc710d3149a064908035
Restrict enabling the manage_snapshot volume feature

The test that is in tempest for this feature is
specific to LVM and will *not* work for other backends
regardless of them supporting the feature. It shouldn't
default to enabled for everyone, only for LVM.

If others want to opt-in they can, but its definitely
the minority that would.

Change-Id: I21347f2a5069059e6413208b254d5acd246faaea