Updated openstack/openstack Project: openstack-infra/project-config  a48b96f5304cdba41558df5ba4a5ac43144821b8
add a command line wrapper for testing the constraints url update

The actual work is done by a function that is called from
make_branch.sh, but that script can't be reused if there's an
issue. This provides a tool for making the updates by hand and also
serves as a way to test changes to the underlying function.

Change-Id: I54b65c3ab9d44e12b4db6c2bbf2f29bcd0488b6b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>