Updated openstack/openstack Project: openstack/neutron  83cfccbfaf21bf3b69d6b7167381dbadbf89e9ca
Only migrate ports on DVR migration

The previous code was always migrating the router ports
if the distributed flag in the router request was True.
This would ultimately turn into a NOOP after the interface
query would return false, but it was a wasted query.

After this patch, it will now only migrate the interface
owners if there is an actual migration to DVR occuring.

TrivialFix
Change-Id: Ia88881dfbd13d735462e9237519aa4bd3e594cc3