Updated openstack/openstack Project: openstack/ironic  5ef20fe6c37b54f6b489259bcd2ce2dae80d9e22
Validate the generated swift temp url

During deployment, a swift temporary URL may be generated,
to be used for getting the image for an instance.

This validates the generated URL to make sure that it
can be accessed.
If validation fails, an exception is raised and the error
is logged. This is helpful when diagnosing the problem with
a deployment.

Change-Id: I1f62bc9b5646d173ec6a3beb6282a5e6002c2030
Closes-Bug: 1512186