Updated openstack/openstack Project: openstack/python-designateclient  7818faba8bcb88d7e39a710991bee6586d6e1f13
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-designateclient.

Change-Id: I2976e5615bd46449690072b09aeaad59c1674d44