Updated openstack/openstack Project: openstack/keystone  320223ed90c2b6f35ca00a2a64e2f61cb1325178
switch @hybrid_property to @property

The purpose of hybrid is so that you can use your attribute at the
class level in a query. At the instance level, it is identical to
@property

A hybrid property is only needed when referenced at the class level,
for instance.

class Wozzle(...)

  @hybrid_property
  def foo(..)

If referencing Wozzle.foo, then you need a hybrid property, otherwise
a regular @property will work just fine.

Change-Id: Ifd3ff9a888919e09b80e87282af829bc14693369
Closes-Bug: 1602422